requirements.txt
transformers
torch
pillow
timm
numpy
opencv-python
gradio